Adding back changes from Limba
Limba had originally submitted this change, along with a big
reformatting change. We reverted the commit to separate the reformatting
from the changes. This commit has a change to the board of directors
formatting and info to contact the management company in the BOD
section. Limba's original commit also had a change to the NCC and the
grounds committee, but we need to gather more info about those changes
before committing.