Changed grounds committee to Ganesh
Even though we don't have Ganesh's contact info, Limba wants to go ahead
and change the contact to be his name. So, the emails will go to Marco,
but people will think they're talking to Ganesh. We'll change it for now
and revert the change if we aren't able to get Ganesh's contact info.